Move ingest consumer around, dropping `v2` suffix
Just a bit of code reorganization
