+++ pyproject.toml
@@ -424,9 +424,7 @@ module = [
     "sentry.incidents.models",
     "sentry.incidents.subscription_processor",
     "sentry.incidents.tasks",
-    "sentry.ingest.consumer_v2.factory",
     "sentry.ingest.inbound_filters",
-    "sentry.ingest.ingest_consumer",
     "sentry.integrations.aws_lambda.integration",
     "sentry.integrations.base",
     "sentry.integrations.bitbucket.client",
