+++ src/sentry/ingest/consumer/attachment_event.py
@@ -6,15 +6,16 @@ import msgpack
 from arroyo.backends.kafka.consumer import KafkaPayload
 from arroyo.types import Message
 
-from sentry.ingest.ingest_consumer import (
+from sentry.models import Project
+from sentry.utils import metrics
+
+from .processors import (
     IngestMessage,
     process_attachment_chunk,
     process_event,
     process_individual_attachment,
     process_userreport,
 )
-from sentry.models import Project
-from sentry.utils import metrics
 
 logger = logging.getLogger(__name__)
