+++ src/sentry/lang/java/utils.py
@@ -7,7 +7,7 @@ import sentry_sdk
 from symbolic.proguard import ProguardMapper
 
 from sentry.attachments import CachedAttachment, attachment_cache
-from sentry.ingest.ingest_consumer import CACHE_TIMEOUT
+from sentry.ingest.consumer.processors import CACHE_TIMEOUT
 from sentry.models import Project, ProjectDebugFile
 from sentry.utils import json
 from sentry.utils.cache import cache_key_for_event
