+++ src/sentry/reprocessing2.py
@@ -219,7 +219,7 @@ def pull_event_data(project_id, event_id) -> ReprocessableEvent:
 
 def reprocess_event(project_id, event_id, start_time):
 
-    from sentry.ingest.ingest_consumer import CACHE_TIMEOUT
+    from sentry.ingest.consumer.processors import CACHE_TIMEOUT
     from sentry.tasks.store import preprocess_event_from_reprocessing
 
     reprocessable_event = pull_event_data(project_id, event_id)
