+++ src/sentry/testutils/pytest/kafka.py
@@ -129,7 +129,7 @@ def session_ingest_consumer(scope_consumers, kafka_admin, task_runner):
     """
 
     def ingest_consumer(settings):
-        from sentry.ingest.consumer_v2.factory import get_ingest_consumer
+        from sentry.ingest.consumer.factory import get_ingest_consumer
         from sentry.ingest.types import ConsumerType
         from sentry.utils.batching_kafka_consumer import create_topics
