+++ tests/sentry/ingest/ingest_consumer/test_ingest_consumer_kafka.py
@@ -10,7 +10,7 @@ from django.conf import settings
 
 from sentry import eventstore
 from sentry.event_manager import EventManager
-from sentry.ingest.consumer_v2.factory import get_ingest_consumer
+from sentry.ingest.consumer.factory import get_ingest_consumer
 from sentry.ingest.types import ConsumerType
 from sentry.testutils.pytest.fixtures import django_db_all
 from sentry.utils import json
